requirements.txt
cycler==0.10.0
decorator==4.0.11
docopt==0.6.2
imageio==2.1.2
numpy==1.13.1
picamera==1.13
Pillow==4.2.1
prompt-toolkit==1.0.15
python-dateutil==2.6.1
pytz==2017.2
RPi.GPIO==0.6.3
simplegeneric==0.8.1
six==1.11.0
tqdm==4.11.2
traitlets==4.3.2
typing==3.6.2
wcwidth==0.1.7
